Update requirements (#215)
**Context:**
Dependencies have not been revisited in a while.

**Description of the Change:**
Updates version of dependencies. In specific, tensorflow is pinned to
version `2.10.1` since version `2.11.0` has changes in the functions
`stack` and `concat` that break the integration with Mr Mustard.

To unpin it we will have to revisit the new tf api for those functions
so that users installing MrMustard with `pip install mrmustard` won't
run into unexpected errors.

Also, since `ray` is an optional dependency, tests related to ray are
now skipped if the package is not installed.

**Benefits:**
Fixes a lot of security issues in TF as alerted by dependabot and also
provides updated versions of the packages.

**Possible Drawbacks:**
TF is pinned — we should address this for the next release. Also, we
should consider moving to a [PEP 621](https://peps.python.org/pep-0621/)
compliant packaging system by either using a `pyproject.toml` file for
dependency specification or a dependency manager like poetry.

**Related GitHub Issues:**
None

---------

Co-authored-by: ziofil <ziofil@users.noreply.github.com>
Co-authored-by: ziofil <miatto@gmail.com>