Improve callbacks and adds tensorboard (#219)
**Context:**
To facilitate the investigation of the optimization process on the
numerics side, as well as giving users more control over the
optimization progress, the callback functionality is expanded to allow
for realtime monitoring and updates. A builtin callback that tracks
optimization progresses in Tensorboard is also provided.

**Description of the Change:**

- The `callback` argument to `Optimizer.minimize` is now `callbacks`
with the new possibility of passing multiple callbacks.
- Adds tagged parameter traversal `Parametrized.traverse_trainables()`
which accumulates the path of tags while traversing. This is now used by
the optimizer to get unique human interpretable identifiers to
parameters (for better tracking).
- Callback functions now accept as argument the optimizer, the cost
value, the cost function, and the trainable parameter values as well as
their gradient. This gives users a lot of run time control, but it's not
necessary to use them all.
- The execution of callbacks moved to after the backprop but before the
parameter update (originally at the end of the step).
- A `Callback` base class for users to extend and inherit from.
- A `TensorboardCallbacks` that does automatic tracking of costs and
parameters (values and gradients) in real time.

Misc changes:

- I added a `path` item to the testing workflow so that non-code changes
won't trigger it.

**Benefits:**
- Better observability.
- Finer control.
- Opens up a lot of possibilities for optimization.

**Possible Drawbacks:**
Users could introduce too much overhead if the callbacks are too heavy,
thus potentially slowing down the optimization.

**Related GitHub Issues:**
None

---------

Co-authored-by: ziofil <ziofil@users.noreply.github.com>