Ensuring that hbar can only be specified in `Settings` (#273)
**Context:**
While the value of `hbar` is specified in `Settings`, some methods allow
passing (potentially different) values of `hbar`. There is no mechanism
to keep track of the different values of `hbar` employed in a
simulation, nor to ensure consistency. This leaves room for unintended
behaviors.
 
**Description of the Change:**
All methods retrieve the value of `hbar` from the `Settings`, instead of
allowing users to specify it in the inputs.

**Benefits:**
The value of `hbar` is guaranteed to be consistent throughout the
session, meaning that at any point, it is equal to the value that can be
found in the `Settings`.

**Possible Drawbacks:**
Users may still change the `Settings` (and specifically, the value of
`hbar`) at any point during the computation. This will be disallowed in
a follow-up Pull Request.

**Related GitHub Issues:**