Freezing settings after their values are used (#278)
**Context:** 
Some of the attributes of `Settings` (e.g., `settings.HBAR`) should
become immutable after they are used for the first time, while others
(such as the cutoffs) can change at any point during the session.

**Description of the Change:**
A new class `ImmutableSetting` is introduced. This class takes care of
"freezing" the value of certain settings after they are queried for the
first time.
Additionally, the `Settings` class is restructured so that every
attribute appears in the docs and is properly documented.

**Remarks:**
We should have discussions about:
- What attributes should become immutable.
- If attributes should that are not currently used, e.g.
`settings.CHOI_R`, should be removed.