Constants and variables (#298)
**Context:**
Improving variables handling, in preparation for the backend refactoring
and the Fock-Bargman refactoring

**Description of the Change:**
- Removed `parameters.py` and `parametrized.py`, with all their
functions and methods. Gates, states, and detectors no longer inherit
from `Parametrized`.
- Added `Constant` and `Variable` classes to represent constants and
variables, and `ParameterSet` to represent sets of parameters. Gates,
states, and detectors own `ParameterSet`s and have fast access to
constant and variable parameters.

**NOT PART OF THIS PR:**
Improvements to the training package (not priority)