Update docs about installing Julia (#290)
**Context:** Julia dependencies (added in order to stabilize the
recurrence relation: https://github.com/XanaduAI/MrMustard/pull/274)

**Description of the Change:** Added Manifest.toml and Project.toml
which specify the julia version and dependencies. Updated the README and
development guide to instruct the user to install julia and to
instantiate the toml files. This means PyJulia now runs in the
environment defined by the toml files, rather than in Julia's base
environment.

- When installing via devcontainer, julia (v1.9.3) is installed and the
environment is instantiaded.
- When installing via MakeFile, the environment is only instantiaded
when Julia was previously installed by the user.
- When installing via `poetry install` or `pip install`, the julia
installation and environment instantiation need to be performed by the
user.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Filippo Miatto <ziofil@users.noreply.github.com>
Co-authored-by: Tanner Rogalsky <tanner@tannerrogalsky.com>
Co-authored-by: ziofil <miatto@gmail.com>