Simplifying tests for compact fock file (#313)
**Context:**
The tests in `tests/test_math/test_compactFock.py` take a long time to
run (over 5mins with the tensorflow backend) and are very intensive in
terms of memory. They often crash on GitHub, making our entire workflow
unstable.

**Description of the Change:**
This PR simplifies `tests/test_math/test_compactFock.py` so that the
tests in it are less intensive in time (1min50s) and memory, but still
test all that they were originally testing.