Adding ``__matmul__`` to ``CircuitComponent``s (#347)
**Context:**
Adds matmul to circuit components.
Also changes/adds a bunch of stuff that is needed for it, such as:
- Adds equality on wires and circuit components (so that tests become
easier)
- Changes ``Wires.__rshift__`` into ``Wires.__matmul__``, so that all
rshift (matmul) in lab_dev mean the same thing, i.e. that they do (do
not) add missing adjoints

**Tests:**
For the tests, it applies ``@`` to sequences of states and gates, and
compares the representation of the resulting component to those that one
gets by implementing the same sequence with `mrmustard.lab`. Small
differences found, which can be discussed offline

---------

Co-authored-by: Filippo Miatto <ziofil@users.noreply.github.com>
Co-authored-by: zy <48225584+zeyueN@users.noreply.github.com>
Co-authored-by: Yuan <16817699+sylviemonet@users.noreply.github.com>
Co-authored-by: Filippo Miatto <miatto@gmail.com>
Co-authored-by: Sebastián Duque Mesa <675763+sduquemesa@users.noreply.github.com>
Co-authored-by: Yuan YAO <sylvieyao502@gmail.com>