Fix the computation of symplectic eigenvalues
The failure listed below was due to the symplectic eigenvalue code
randomly returning the same symplectic eigenvalue multiple times.  This
occured because it was assumed that the computed eigenvalues of the
operator |iΩΣ| were sorted.  This assumption is not true for all cases.

This fix ensures that the eigenvalues of the operator are sorted so that
the assumption is true and the symplectic eigenvalues are indeed every
second entry.  The test_random_state_is_entangled test always passes
with this fix.

```
================================== FAILURES ===================================
_______________________ test_random_state_is_entangled ________________________

    def test_random_state_is_entangled():
        """Tests that a Gaussian state generated at random is entangled."""
        state = Vacuum(2) >> Ggate(num_modes=2)
        mat = state.cov
        assert np.allclose(gp.log_negativity(mat), 0.0)
        assert np.allclose(
            gp.log_negativity(gp.physical_partial_transpose(mat, [0, 1])), 0.0, atol=1e-7
        )
        N1 = gp.log_negativity(gp.physical_partial_transpose(mat, [0]))
        N2 = gp.log_negativity(gp.physical_partial_transpose(mat, [1]))

>       assert N1 > 0
E       assert 0 > 0

tests/test_lab/test_states.py:223: AssertionError
```