enable new TF and pip installation (#406)
**Context:**
If you install MrMustard with `pip`, it will install the latest
tensorflow, but that does not allow the use of
`tf.keras.optimizers.legacy.Adam`. We need the non-legacy version, but
it's [slow on M1/M2
Macs](https://stackoverflow.com/questions/77222463/is-there-a-way-to-change-adam-to-legacy-when-using-mac-m1-m2-in-tensorflow).
Some manual effort is required to support the legacy optimizer with TF
2.16+, and it seems out of the door anyway. I'd advise Mac M1/M2 users
to install TF 2.15 in advance if they want the better optimizer for them
(until TF fixes it, at least).

**Description of the Change:**
- Add the `tf` extra to `tensorflow-probability`. This one's a bit
hacky, because that extra doesn't exist until 0.24+, but if you are
using 0.24+ (which you would be with a `pip install`), you need it to
use Keras optimizers. The good news is that `poetry` doesn't seem to
mind, and totally ignores it for our lockfile (which has it pinned to
0.22)
- Use the non-legacy optimizer for TF 2.16+ since it will work out of
the box
- Raise a warning to advise M1/M2 Mac users to downgrade to TF 2.15 (if
they're on 2.16+)

**Benefits:**
If users pip-install MrMustard, it will still work (even with TF 2.16)

**Possible Drawbacks:**
Mac M1/M2 users will have slow optimizations if they get a fresh
environment and `pip install mrmustard`. They can avoid this by
pre-installing TF 2.15 to their environment. I've added a warning to
tell them clearly what to do.

**Related GitHub Issues:**
Fixes #381 by adding 2.16 support