Operator and Map [merge before QtoB] (#405)
**Context:**
The `Unitary` class isn't enough, because we have things that are
operators but not unitary. So we introduce `Operator` and `Map` to stay
between `CircuitComponent` and `Unitary` and `Channel`, respectively.

**Description of the Change:**
- Added code of `Operator` and `Map`, updated names of `BtoQMap` to
`BtoQ` and now it inherits from `Operator` (so no need for "map" in the
name). Also `DsMap` now becomes `BtoPS` and it inherits from `Map`.
- Moved the `name` field to the end in several methods.

**Benefits:**
Clearer structure, less mistakes going forward, getting things ready for
the `MultiRepresentation` class.

---------

Co-authored-by: Matthew Silverman <ma.silv11@gmail.com>