rshift of Ket and DM returns scalar if no Hilbert spaces (#411)
Right-shifting two components so that no wires are left, like
`Coherent([0], 1.0) >> Vacuum([0]).dual`, now returns a scalar, which as
Jacob points out is more intuitive.

This is a tricky thing to do in code because it's easier if the return
type of a method (`__rshift__` here) is always the same and not
dependent on context. However, I think I did things carefully and I kept
the behaviour confined to the `>>` operator: `Coherent([0], 1.0) @
Vacuum([0]).dual` returns a components with no wires and scalar
representation. This way we can still use `@` internally and not worry
about scalars being introduced.

There is one (possibly rare) edge case that is not possible to handle:
things like `Coherent([0], 1.0) >> Vacuum([0]).dual >> Coherent([0],
1.0)`, which should just return a rescaled coherent state now gives a
nonsensical result, because the `__rshift__` of the array resulting from
the first `>>` gets called first and we cannot avoid that. If the scalar
comes _after_ the circuit component then it works.

---------

Co-authored-by: Matthew Silverman <ma.silv11@gmail.com>
Co-authored-by: Eli Bourassa <53090166+elib20@users.noreply.github.com>