Autoshape and fock improvements (#397)
**Context:**
A recommended Fock shape of the components of a circuit can often be
determined by the nature of the components and their connectivity. Also,
for certain components like `States` we can determine a good choice of
shape by stopping e.g. at 99.9% of the norm.

**Description of the Change:**
- `CircuitComponent`s now have the property `.manual_shape` which can be
set manually. It starts as `[None,...]` and it can be modified like e.g.
`my_state.manual_shape[1]=9` and now it will be `[None, 9, None, ...]`.
- `CircuitComponent`s now have the method `.auto_shape()` which computes
and returns a tuple with a guesstimate of the recommended shape to the
best of our knowledge. It will respect the values in `.manual_shape`
that have already been set.
- `Circuit` now has the method `propagate_shapes()` which exploits the
context of each component to produce a more compact shape than
`auto_shape()` would give in isolation. This gives the `Circuit` the
power to contract an arrangement of components in the `Fock`
representation in a close to optimal way (excluding throwing SVDs
around), when combined with the optimal contraction path (upcoming).
- Faster `autocutoff` method (now `autoshape_numba`). It doesn't any
longer travel back and forth between phase space and Bargmann for each
mode, and it's the same as Robbe's diagonal method, but specialized for
this task, so it's faster.
- General "fock-related" functionality has been refactored/improved
    - removed the converter module
- localized the fock conversion into the `CircuitComponent` so that wire
information is available
- components keep the original bargmann data when they are converted to
fock
    - can no longer matmul across representations (this is good)

---------

Co-authored-by: Matthew Silverman <ma.silv11@gmail.com>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: Eli Bourassa <53090166+elib20@users.noreply.github.com>