improve visualizations in lab_dev (#425)
**Context:**
There have been some issues with visualizations in MrMustard, and this
should address a good amount of them

**Description of the Change:**
- Remove mako-based implementations for `_repr_html_`, and replace them
with ipywidgets-based implementations with `_ipython_display_`. I mainly
did this because ipywidgets is easier to work with than raw HTML, and
should be more portable (eg. jupyterlab vs VS Code). Also, the switch to
`_ipython_display_` allows for side-effects, and doesn't also print the
default repr (unless I do it explicitly)
- Simplified the funky HTML+CSS to a consistent and basic look. Easier
to update, and not distracting at all.
- fix the default `return_fig=False` option for `State.visualize_*()`
methods by not converting to HTML manually
- some little things to please codefactor

**Benefits:**
- Works everywhere!
- Easier to maintain
- Uses highly trusted ipywidgets package
- No more double-repr

**Possible Drawbacks:**
- N/A

**Related GitHub Issues:**
Fixes #423

---------

Co-authored-by: Anthony <125415978+apchytr@users.noreply.github.com>