Call none (#439)
**Context:**
Adding a method where one can call the ansatz with `None` for some
arguments. This will return a an updated ansatz with on fewer variables
(wires).
**Description of the Change:**
Adding a method to the `DiffOpPolyExpAnsatz` where one can evaluate only
some of the wires, and thus returning a new updated ansatz.
**Benefits:**
Allows for reduction in the size of the ansatz. Can also be used with
`BtoQ` to do a Homodyne measurement.

---------

Co-authored-by: Filippo Miatto <ziofil@users.noreply.github.com>
Co-authored-by: Eli Bourassa <53090166+elib20@users.noreply.github.com>
Co-authored-by: Anthony <125415978+apchytr@users.noreply.github.com>