Adding classical wires (#456)
**Context:** It is apparent that Mr Mustard requires some notion of
'classical wire' to support ~~both the new poly exp ansatz as well as~~
measurements. Here, I introduce `classical_output` and `classical_input`
(or some other name we can agree upon) into `Wires` that is meant to
fill this role. The idea is that `Wires` will keep track of the modes
that contain this classical information (associating it with an index
and unique ID) and the representation will do the rest (in a followup
PR).

**Description of the Change:** `Wires` now accepts `classical_output`
and `classical_input` meant to fill the role of non-quantum wires. The
behaviour (within Wires) is identical to bra and ket.

**Benefits:** We will now be able to link classical information between
circuit components. This approach avoids the need to introduce several
new `Wires` classes.

**Possible Drawbacks:** It's possible there are some use cases that may
require us to reconsider this approach? ~~Maybe it turns out we really
do need separate classes (i.e `ClassicalWire`, `QuantumWire`,
`PolyWire`).~~