Serialization (#451)
**Context:**
For execution in certain environments, it helps to be able to
de/serialize objects

**Description of the Change:**
- Introduce the `utils.serialize` module for saving and loading
MrMustard objects that implement the interface
- Implement serialization helpers for `CircuitComponent`, `Bargmann` and
`Fock` (and `Number` to avoid a tensorflow bug)
- Implement de/serialize for `Circuit`, consuming info organized by
those helpers then writing to 2 (zipped) files
- 🐛 fix: mode order in `Operation` constructor is now correct
- 🐛 fix: `ParameterSet.all_parameters` no longer adds the variables to
the constants

***Implementation details***
- The Representations return a dict mapping a variable name to numpy/tf
data
- The `CircuitComponent` one returns two items, serializable kwargs and
non-serializable data
- If `name` is present in the component type `__init__` method, it
serializes using values needed for `_from_attributes`. This is never the
case for concrete classes, but a backup if any class does provide it
- Otherwise, the serializer tries to be clever by checking if `modes` or
`modes_in + modes_out` are present. If neither are, it will fail. Then
it passes all params from the `ParameterSet`, adding `<p>_trainable` and
`<p>_bounds` if it's a variable as opposed to a constant
- When calling `save` without `arrays`, it creates a `.json` file. If
you provide `arrays` (which `Circuit.serialize` does), it creates a
`.zip` file with two files in it: the regular JSON file, and a `.npz`
file that contains the associated numpy data. This helps with
portability. If you'd rather not zip things (eg. perf over portability),
you can pass `do_zip=False` to `save`

**Benefits:**
Useful for certain execution environments

**Possible Drawbacks:**
- I assume that the `ParameterSet` contents are enough to serialize.
This was not true for the maps, so I updated their parameter sets
- I assume that `modes_in` and `modes_out` imply that `in_ket ==
out_ket`, and that `in_bra == out_bra`
- I assume that any class constructed with `modes` will have all
in/out/bra/ket modes equal or empty

**Related GitHub Issues:**
N/A

---------

Co-authored-by: Anthony <125415978+apchytr@users.noreply.github.com>