Fock to bargmann (#457)
**Context:**
Adding a method to convert circuit components from Fock to Bargmann
**Description of the Change:**
When contracting a bargmann and a fock object with each other, the fock
object is converted to bargmann. This is opposite of what is currently
done. If two fock objects meet each other they stay in fock.
**Benefits:**
- Allows all the benefits from the new ansatz without changing the API.
(i.e. `Number.dual` as a measurement device).
- All objects in fock will automatically be converted to bargmann, which
is more efficient for contracting with other objects down the line. It
will create a "identity" Bargmann object with the Fock coefficients
attached as the array (c)

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Filippo Miatto <ziofil@users.noreply.github.com>
Co-authored-by: Eli Bourassa <53090166+elib20@users.noreply.github.com>
Co-authored-by: Anthony <125415978+apchytr@users.noreply.github.com>