X and Y of a channel (#470)
**Context:**
Added the ability to compute X and Y of a given channel.

**Description of the Change:**
Added `XY_of_channel` to the physics/bargmann.py that carry out the
computation. It transforms the A matrix of a channel to the
corresponding X and Y matrices. Moreover, added the property to channels
that can be used by `ch.XY` for any instance `ch` of a Channel object.

---------

Co-authored-by: Filippo Miatto <ziofil@users.noreply.github.com>