add juliapkg.json to package data (#482)
**Context:**
If you installed from a clone of the repo, you'll have `juliapkg.json`
present. If you install from PyPI... not the case. This means that if a
dependent project (eg. MrKite) tries to use MrMustard's Julia features,
the Julia deps will not be auto-managed because this file won't exist in
that venv!

**Description of the Change:**
~Add `juliapkg.json` to `tool.setuptools.package-data` so dependent
projects use Julia correctly.~ (`poetry` did not respect the setuptools
thing easily, and we don't need it because moving the file works)

Move `juliapkg.json` into the `mrmustard/` folder so it's installed with
`mrmustard`. Confirmed that `juliacall` finds it both with editable
installs of MrMustard, and with it added to another project (eg. MrKite)
as a dependency.

**Benefits:**
Places that install `MrMustard` from PyPI can use the Julia features

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
XanaduAI/MrKite#196