Sampler (#472)
**Context:** As part of measurement devices in lab_dev we require a
sampler class. The sampler can be though of as a black box that takes in
a state and outputs either a probability distribution or a collection of
samples. Here I implement said class along with subclasses for PNR
samples and Homodyne samples where the sampler is designed in way to
easily allow for performance shortcuts. Note: some implementation
details are subject to change as we continue to develop measurement
devices.

**Description of the Change:** `Sampler`, `PNRSampler`,
`HomodyneSampler`.