Bump `rich` dependency to version 13
**Context:** Old dependencies can hold back things which depend on the
MrMustard, which can be very deep in the dependency tree.  The `rich`
package had a very old major version dependency (10.*.*) and this could
interact with other seemingly random packages when depending on
MrMustard.

**Description of the Change:** Bump the major version of `rich` to 13.
The only place I can see to test this version change is in the
ProgressBar class, and this seemed to work fine for me on testing using
the example code in the docs for using the optimizer.

**Benefits:** Other packages that depend on MrMustard won't get into a
tricky dependency situation with a very old package.

**Possible Drawbacks:** Possible subtle bugs as this dependency doesn't
appear to have unit tests.