An initial implementation of the phase noise channel (#513)
**Context:**
Adds the phase noise with a Fock representation.

**Description of the Change:**
Added a class under `Channel`. Please note that if we wanted to actually
write a Fock array for this channel, it would be extremely heavy (I ran
into memory issues). Therefore, we implement contractions with
`PhaseNoise` in a particular way, using a custum rrshift method. I
believe a similar logic was used in lab before.

**Benefits:**
Enables us to apply phase noise.

**Possible Drawbacks:**
Could be slow.