Getting rid of CircuitComponent level access of parameters in the ParameterSet (#534)
**Context:** By exposing parameters in the parameter set of a circuit
component we introduce the potential for name collisions. In fact, we
already see this with `symplectic` on `Ggate`. The alternative is to
access parameters only through the parameter set itself.

**Description of the Change:** Removed `CircuitComponent._add_parameter`
no longer exposing parameters at the `CircuitComponent` level. Renamed
the attribute `parameter_set` to `parameters`. Fixed the inits of states
and transformations to properly handle parameters. Added the
`symplectics.py` file to `physics`.

**Benefits:** Fewer attribute name collisions. Parameters are properly
handled now.

**Possible Drawbacks:** Minor inconvenience having to specify
`.parameters`