Remove `julia` code and dependencies (#541)
Opening on behalf of @aplund 

**Context:** Julia code was introduced previously to try and work around
algorithmic instabilities in the computation of Fock coefficients from
Gaussian Bargmann functions. This introduced a hard dependency on
`julia`, which made installation tricker. In the end, it was found that
the more stable approach was achieved by changing the algorithm to
average over the possible lattice sites.

**Description of the Change:** This change removes the `julia` code and
the dependencies of that code. The PRECISION_BITS_HERMITE_POLY setting
is removed, as that was the switch to turn on the `julia`
implementation. The `julia` code has been pushed into the branch
<https://github.com/XanaduAI/GaussianBargmannFock> which stores
implementations of this function.

**Benefits:** There is no hard dependency on `juliacall` and the `julia`
language. This should ease installation, particularly on Windows
platforms.

**Possible Drawbacks:** The ability switch to using the `julia`
implementation is gone without a setting to bring it back quickly.

---------

Signed-off-by: Anthony <anthonychytros@gmail.com>
Co-authored-by: Austin Lund <austin.lund@xanadu.ai>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>