Forcing different representations to use Fock  (#559)
**Context:** Issues like [MemoryError
#481](https://github.com/XanaduAI/MrMustard/issues/481) are currently
present in develop. These occur when attempting to compute the L2_norm
of a state with a non scalar c. The solution presented here is to force
different representation contractions to occur in Fock and to force the
L2 norm to compute in Fock when the number of derived variables is > 0.

**Description of the Change:** Representation now raises when two
different ansatz are met. Circuit component now handles changing the
representation and uses `to_fock` instead. Computing the L2 is done in
Fock when c is not scalar. Added #481 as a test.

**Benefits:** normalization and purity calculation issues resolved

---------

Co-authored-by: Filippo Miatto <ziofil@users.noreply.github.com>