allow backend hot-swap (#568)
**Context:**
When working in a jupyter notebook, swapping the backend requires one to
fully restart the kernel, which can be very annoying.

**Description of the Change:**
This PR removes that limitation.

**Benefits:**
One can save pre-computed values and not lose the result of computations
when swapping between backends.

**Possible Drawbacks:**
Objects created before the swap are still based on the previous backend

Co-authored-by: Anthony <125415978+apchytr@users.noreply.github.com>