OptimizerJax bugfix: reset opt_history  (#634)
**Context:** An `OptimizerJax` instance should be reusable but currently
is not due to the population of `opt_history` which tricks the optimizer
into thinking that it has finished optimizing on subsequent runs. This
PR introduces a bugfix that allows users to reuse a single
`OptimizerJax` instance

**Description of the Change:** `OptimizerJax.minimize` now clears
`opt_history`. Added `test_reuse_optimizer`. Reorganized the tests in
`test_opt_lab_jax.py` in alphabetical order.

**Benefits:** Can reuse the same Jax instance which potentially allows
us to avoid some recompilation.